refactor: Complete orchestrator refactoring by moving methods to managers
- Move verify_target_capabilities to ExecutionManager
- Move analyze_run, _update_global_coverage, _calculate_coverage_hash,
  _build_lineage_profile to ScoringManager
- Move _calculate_mutations, _get_nodes_from_parent, get_boilerplate,
  _extract_and_cache_boilerplate, _get_core_code to MutationController
- Move _log_timeseries_datapoint, update_and_save_run_stats to ArtifactManager
- Update all call sites in orchestrator to delegate to managers
- Update all tests to use the new manager method locations
