+++ tests/orchestrator/test_main_loop.py
@@ -12,19 +12,20 @@ from datetime import datetime, timezone
 from pathlib import Path
 from unittest.mock import MagicMock, patch
 
+from lafleur.mutation_controller import MutationController
 from lafleur.orchestrator import LafleurOrchestrator
 
 
 class TestCalculateMutations(unittest.TestCase):
-    """Test _calculate_mutations method."""
+    """Test MutationController._calculate_mutations method."""
 
     def setUp(self):
-        self.orchestrator = LafleurOrchestrator.__new__(LafleurOrchestrator)
+        self.controller = MutationController.__new__(MutationController)
 
     def test_base_mutations_with_score_100(self):
         """Test that score of 100 gives base mutations."""
         with patch("sys.stdout", new_callable=io.StringIO):
-            mutations = self.orchestrator._calculate_mutations(100.0)
+            mutations = self.controller._calculate_mutations(100.0)
 
             # Score of 100 should give close to base (100)
             # Multiplier = 0.5 + log(100 * 10 / 100) / 2 = 0.5 + log(10) / 2 ≈ 0.5 + 1.15 = 1.65
@@ -35,7 +36,7 @@ class TestCalculateMutations(unittest.TestCase):
     def test_low_score_clamped_to_minimum(self):
         """Test that very low scores result in minimum multiplier."""
         with patch("sys.stdout", new_callable=io.StringIO):
-            mutations = self.orchestrator._calculate_mutations(1.0)
+            mutations = self.controller._calculate_mutations(1.0)
 
             # Score of 1.0:
             # score_multiplier = 1.0 / 100.0 = 0.01
@@ -47,7 +48,7 @@ class TestCalculateMutations(unittest.TestCase):
     def test_high_score_clamped_to_maximum(self):
         """Test that very high scores are clamped to 3.0x multiplier."""
         with patch("sys.stdout", new_callable=io.StringIO):
-            mutations = self.orchestrator._calculate_mutations(10000.0)
+            mutations = self.controller._calculate_mutations(10000.0)
 
             # Very high score should clamp to 3.0x, so 100 * 3.0 = 300
             self.assertEqual(mutations, 300)
@@ -55,7 +56,7 @@ class TestCalculateMutations(unittest.TestCase):
     def test_medium_score_calculates_dynamic_multiplier(self):
         """Test that medium scores use logarithmic scaling."""
         with patch("sys.stdout", new_callable=io.StringIO):
-            mutations = self.orchestrator._calculate_mutations(50.0)
+            mutations = self.controller._calculate_mutations(50.0)
 
             # Score of 50:
             # score_multiplier = 50 / 100 = 0.5
@@ -67,7 +68,7 @@ class TestCalculateMutations(unittest.TestCase):
     def test_prints_dynamic_adjustment_message(self):
         """Test that dynamic adjustment is logged."""
         with patch("sys.stdout", new_callable=io.StringIO) as mock_stdout:
-            self.orchestrator._calculate_mutations(75.0)
+            self.controller._calculate_mutations(75.0)
 
             stdout_output = mock_stdout.getvalue()
             self.assertIn("Dynamically adjusting mutation count", stdout_output)
@@ -75,13 +76,14 @@ class TestCalculateMutations(unittest.TestCase):
 
 
 class TestUpdateAndSaveRunStats(unittest.TestCase):
-    """Test update_and_save_run_stats method."""
+    """Test ArtifactManager.update_and_save_run_stats method."""
 
     def setUp(self):
-        self.orchestrator = LafleurOrchestrator.__new__(LafleurOrchestrator)
-        self.orchestrator.run_stats = {}
-        self.orchestrator.coverage_manager = MagicMock()
-        self.orchestrator.coverage_manager.state = {
+        from lafleur.artifacts import ArtifactManager
+
+        self.run_stats = {}
+        self.coverage_manager = MagicMock()
+        self.coverage_manager.state = {
             "per_file_coverage": {
                 "file1.py": {},
                 "file2.py": {},
@@ -93,70 +95,82 @@ class TestUpdateAndSaveRunStats(unittest.TestCase):
                 "rare_events": {6: 1},
             },
         }
-        self.orchestrator.global_seed_counter = 42
-        self.orchestrator.corpus_manager = MagicMock()
-        self.orchestrator.corpus_manager.corpus_file_counter = 123
+        self.corpus_manager = MagicMock()
+        self.corpus_manager.corpus_file_counter = 123
+
+        # Create ArtifactManager with required dependencies
+        self.artifact_manager = ArtifactManager.__new__(ArtifactManager)
+        self.artifact_manager.run_stats = self.run_stats
+        self.artifact_manager.coverage_manager = self.coverage_manager
+        self.artifact_manager.corpus_manager = self.corpus_manager
 
     def test_updates_timestamp(self):
         """Test that last_update_time is set to current time."""
-        with patch("lafleur.orchestrator.save_run_stats"):
-            before = datetime.now(timezone.utc)
-            self.orchestrator.update_and_save_run_stats()
-            after = datetime.now(timezone.utc)
+        with patch("lafleur.artifacts.save_run_stats"):
+            with patch("lafleur.artifacts.generate_corpus_stats"):
+                before = datetime.now(timezone.utc)
+                self.artifact_manager.update_and_save_run_stats(global_seed_counter=42)
+                after = datetime.now(timezone.utc)
 
-            timestamp = datetime.fromisoformat(self.orchestrator.run_stats["last_update_time"])
-            self.assertGreaterEqual(timestamp, before)
-            self.assertLessEqual(timestamp, after)
+                timestamp = datetime.fromisoformat(self.run_stats["last_update_time"])
+                self.assertGreaterEqual(timestamp, before)
+                self.assertLessEqual(timestamp, after)
 
     def test_updates_corpus_size(self):
         """Test that corpus_size reflects per_file_coverage count."""
-        with patch("lafleur.orchestrator.save_run_stats"):
-            self.orchestrator.update_and_save_run_stats()
+        with patch("lafleur.artifacts.save_run_stats"):
+            with patch("lafleur.artifacts.generate_corpus_stats"):
+                self.artifact_manager.update_and_save_run_stats(global_seed_counter=42)
 
-            self.assertEqual(self.orchestrator.run_stats["corpus_size"], 3)
+                self.assertEqual(self.run_stats["corpus_size"], 3)
 
     def test_updates_global_coverage_counts(self):
         """Test that global coverage metrics are counted."""
-        with patch("lafleur.orchestrator.save_run_stats"):
-            self.orchestrator.update_and_save_run_stats()
+        with patch("lafleur.artifacts.save_run_stats"):
+            with patch("lafleur.artifacts.generate_corpus_stats"):
+                self.artifact_manager.update_and_save_run_stats(global_seed_counter=42)
 
-            self.assertEqual(self.orchestrator.run_stats["global_uops"], 2)
-            self.assertEqual(self.orchestrator.run_stats["global_edges"], 3)
-            self.assertEqual(self.orchestrator.run_stats["global_rare_events"], 1)
+                self.assertEqual(self.run_stats["global_uops"], 2)
+                self.assertEqual(self.run_stats["global_edges"], 3)
+                self.assertEqual(self.run_stats["global_rare_events"], 1)
 
     def test_updates_seed_counters(self):
         """Test that global_seed_counter and corpus_file_counter are updated."""
-        with patch("lafleur.orchestrator.save_run_stats"):
-            self.orchestrator.update_and_save_run_stats()
+        with patch("lafleur.artifacts.save_run_stats"):
+            with patch("lafleur.artifacts.generate_corpus_stats"):
+                self.artifact_manager.update_and_save_run_stats(global_seed_counter=42)
 
-            self.assertEqual(self.orchestrator.run_stats["global_seed_counter"], 42)
-            self.assertEqual(self.orchestrator.run_stats["corpus_file_counter"], 123)
+                self.assertEqual(self.run_stats["global_seed_counter"], 42)
+                self.assertEqual(self.run_stats["corpus_file_counter"], 123)
 
     def test_calculates_average_mutations_per_find(self):
         """Test that average_mutations_per_find is calculated when finds > 0."""
-        self.orchestrator.run_stats["new_coverage_finds"] = 10
-        self.orchestrator.run_stats["sum_of_mutations_per_find"] = 250
+        self.run_stats["new_coverage_finds"] = 10
+        self.run_stats["sum_of_mutations_per_find"] = 250
 
-        with patch("lafleur.orchestrator.save_run_stats"):
-            self.orchestrator.update_and_save_run_stats()
+        with patch("lafleur.artifacts.save_run_stats"):
+            with patch("lafleur.artifacts.generate_corpus_stats"):
+                self.artifact_manager.update_and_save_run_stats(global_seed_counter=42)
 
-            self.assertEqual(self.orchestrator.run_stats["average_mutations_per_find"], 25.0)
+                self.assertEqual(self.run_stats["average_mutations_per_find"], 25.0)
 
     def test_no_average_when_no_finds(self):
         """Test that average_mutations_per_find is not set when no finds."""
-        self.orchestrator.run_stats["new_coverage_finds"] = 0
+        self.run_stats["new_coverage_finds"] = 0
 
-        with patch("lafleur.orchestrator.save_run_stats"):
-            self.orchestrator.update_and_save_run_stats()
+        with patch("lafleur.artifacts.save_run_stats"):
+            with patch("lafleur.artifacts.generate_corpus_stats"):
+                self.artifact_manager.update_and_save_run_stats(global_seed_counter=42)
 
-            self.assertNotIn("average_mutations_per_find", self.orchestrator.run_stats)
+                self.assertNotIn("average_mutations_per_find", self.run_stats)
 
     def test_calls_save_run_stats(self):
         """Test that save_run_stats is called with updated stats."""
-        with patch("lafleur.orchestrator.save_run_stats") as mock_save:
-            self.orchestrator.update_and_save_run_stats()
+        with patch("lafleur.artifacts.save_run_stats") as mock_save:
+            with patch("lafleur.artifacts.generate_corpus_stats"):
+                self.artifact_manager.update_and_save_run_stats(global_seed_counter=42)
 
-            mock_save.assert_called_once_with(self.orchestrator.run_stats)
+                mock_save.assert_called_once_with(self.run_stats)
 
 
 class TestRunEvolutionaryLoop(unittest.TestCase):
@@ -174,9 +188,8 @@ class TestRunEvolutionaryLoop(unittest.TestCase):
         self.orchestrator.deepening_probability = 0.3
         self.orchestrator.score_tracker = MagicMock()
         self.orchestrator.global_seed_counter = 0
-        self.orchestrator.analyze_run = MagicMock()
-        self.orchestrator._build_lineage_profile = MagicMock()
-        self.orchestrator._log_timeseries_datapoint = MagicMock()
+        self.orchestrator.artifact_manager = MagicMock()
+        self.orchestrator.scoring_manager = MagicMock()
 
     def test_bootstraps_corpus_when_below_minimum(self):
         """Test that corpus generation is triggered when size < min."""
@@ -190,16 +203,15 @@ class TestRunEvolutionaryLoop(unittest.TestCase):
         # Mock to break the infinite loop after bootstrap
         with patch.object(self.orchestrator, "execute_mutation_and_analysis_cycle"):
             with patch("sys.stderr", new_callable=io.StringIO):
-                with patch("lafleur.orchestrator.save_run_stats"):
-                    with patch.object(
-                        self.orchestrator.corpus_manager, "select_parent", return_value=None
-                    ):
-                        self.orchestrator.run_evolutionary_loop()
+                with patch.object(
+                    self.orchestrator.corpus_manager, "select_parent", return_value=None
+                ):
+                    self.orchestrator.run_evolutionary_loop()
 
-                        # Should generate 3 files
-                        self.assertEqual(
-                            self.orchestrator.corpus_manager.generate_new_seed.call_count, 3
-                        )
+                    # Should generate 3 files
+                    self.assertEqual(
+                        self.orchestrator.corpus_manager.generate_new_seed.call_count, 3
+                    )
 
     def test_warns_when_fusil_path_invalid(self):
         """Test that warning is printed when fusil path is invalid."""
@@ -209,14 +221,13 @@ class TestRunEvolutionaryLoop(unittest.TestCase):
         self.orchestrator.corpus_manager.select_parent = MagicMock(return_value=None)
 
         with patch("sys.stderr", new_callable=io.StringIO) as mock_stderr:
-            with patch("lafleur.orchestrator.save_run_stats"):
-                with patch("sys.exit") as mock_exit:
-                    self.orchestrator.run_evolutionary_loop()
+            with patch("sys.exit") as mock_exit:
+                self.orchestrator.run_evolutionary_loop()
 
-                    stderr_output = mock_stderr.getvalue()
-                    self.assertIn("WARNING: Cannot generate new seed files", stderr_output)
-                    self.assertIn("/invalid/path/to/fusil", stderr_output)
-                    mock_exit.assert_called_once_with(1)
+                stderr_output = mock_stderr.getvalue()
+                self.assertIn("WARNING: Cannot generate new seed files", stderr_output)
+                self.assertIn("/invalid/path/to/fusil", stderr_output)
+                mock_exit.assert_called_once_with(1)
 
     def test_exits_when_corpus_empty_and_no_seeder(self):
         """Test that sys.exit is called when corpus is empty with no seeder."""
@@ -225,13 +236,12 @@ class TestRunEvolutionaryLoop(unittest.TestCase):
         self.orchestrator.corpus_manager.select_parent = MagicMock(return_value=None)
 
         with patch("sys.stderr", new_callable=io.StringIO) as mock_stderr:
-            with patch("lafleur.orchestrator.save_run_stats"):
-                with patch("sys.exit") as mock_exit:
-                    self.orchestrator.run_evolutionary_loop()
+            with patch("sys.exit") as mock_exit:
+                self.orchestrator.run_evolutionary_loop()
 
-                    stderr_output = mock_stderr.getvalue()
-                    self.assertIn("CRITICAL: The corpus is empty", stderr_output)
-                    mock_exit.assert_called_once_with(1)
+                stderr_output = mock_stderr.getvalue()
+                self.assertIn("CRITICAL: The corpus is empty", stderr_output)
+                mock_exit.assert_called_once_with(1)
 
     # def test_proceeds_with_existing_files_when_seeder_unavailable(self):
     #     """Test that loop proceeds when corpus has files but below minimum."""
@@ -320,9 +330,13 @@ class TestExecuteMutationAndAnalysisCycle(unittest.TestCase):
         parent_path = Path("/corpus/parent_test.py")
         parent_score = 150.0
 
-        with patch.object(self.orchestrator, "_calculate_mutations", return_value=200) as mock_calc:
+        with patch.object(
+            self.orchestrator.mutation_controller, "_calculate_mutations", return_value=200
+        ) as mock_calc:
             with patch.object(
-                self.orchestrator, "_get_nodes_from_parent", return_value=(None, None, [])
+                self.orchestrator.mutation_controller,
+                "_get_nodes_from_parent",
+                return_value=(None, None, []),
             ):
                 with patch("sys.stderr", new_callable=io.StringIO):
                     self.orchestrator.execute_mutation_and_analysis_cycle(
@@ -335,9 +349,13 @@ class TestExecuteMutationAndAnalysisCycle(unittest.TestCase):
         """Test that cycle returns when _get_nodes_from_parent fails."""
         parent_path = Path("/corpus/parent_test.py")
 
-        with patch.object(self.orchestrator, "_calculate_mutations", return_value=100):
+        with patch.object(
+            self.orchestrator.mutation_controller, "_calculate_mutations", return_value=100
+        ):
             with patch.object(
-                self.orchestrator, "_get_nodes_from_parent", return_value=(None, None, [])
+                self.orchestrator.mutation_controller,
+                "_get_nodes_from_parent",
+                return_value=(None, None, []),
             ):
                 with patch("sys.stderr", new_callable=io.StringIO):
                     result = self.orchestrator.execute_mutation_and_analysis_cycle(
@@ -354,9 +372,11 @@ class TestExecuteMutationAndAnalysisCycle(unittest.TestCase):
         mock_harness.name = "uop_harness_test"
         mock_tree = MagicMock()
 
-        with patch.object(self.orchestrator, "_calculate_mutations", return_value=2):
+        with patch.object(
+            self.orchestrator.mutation_controller, "_calculate_mutations", return_value=2
+        ):
             with patch.object(
-                self.orchestrator,
+                self.orchestrator.mutation_controller,
                 "_get_nodes_from_parent",
                 return_value=(mock_harness, mock_tree, []),
             ):
@@ -381,9 +401,11 @@ class TestExecuteMutationAndAnalysisCycle(unittest.TestCase):
         mock_harness.name = "uop_harness_test"
         mock_tree = MagicMock()
 
-        with patch.object(self.orchestrator, "_calculate_mutations", return_value=100):
+        with patch.object(
+            self.orchestrator.mutation_controller, "_calculate_mutations", return_value=100
+        ):
             with patch.object(
-                self.orchestrator,
+                self.orchestrator.mutation_controller,
                 "_get_nodes_from_parent",
                 return_value=(mock_harness, mock_tree, []),
             ):
@@ -411,9 +433,11 @@ class TestExecuteMutationAndAnalysisCycle(unittest.TestCase):
         mock_harness.name = "uop_harness_test"
         mock_tree = MagicMock()
 
-        with patch.object(self.orchestrator, "_calculate_mutations", return_value=1):
+        with patch.object(
+            self.orchestrator.mutation_controller, "_calculate_mutations", return_value=1
+        ):
             with patch.object(
-                self.orchestrator,
+                self.orchestrator.mutation_controller,
                 "_get_nodes_from_parent",
                 return_value=(mock_harness, mock_tree, []),
             ):
@@ -446,9 +470,11 @@ class TestExecuteMutationAndAnalysisCycle(unittest.TestCase):
         # Configure execution_manager.execute_child to return (None, None) tuple
         self.orchestrator.execution_manager.execute_child.return_value = (None, None)
 
-        with patch.object(self.orchestrator, "_calculate_mutations", return_value=1):
+        with patch.object(
+            self.orchestrator.mutation_controller, "_calculate_mutations", return_value=1
+        ):
             with patch.object(
-                self.orchestrator,
+                self.orchestrator.mutation_controller,
                 "_get_nodes_from_parent",
                 return_value=(mock_harness, MagicMock(), []),
             ):
